requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.40.0
langsmith>=0.1.0
httpx>=0.27.0